v2実装開始

+++ AmvPlayer/video/src/main/java/com/michael/video/AmvSettings.kt
@@ -1,25 +1,46 @@
 package com.michael.video
 
+import android.content.Context
+import android.content.pm.PackageManager
 import android.os.Build
+import io.github.toyota32k.utils.UtLog
+import okhttp3.OkHttpClient
 import java.io.File
 
+interface IAmvHttpClientSource {
+    fun getHttpClient():OkHttpClient
+}
+
 object AmvSettings {
+    val logger = UtLog("AMV", null, "com.michael.video.")
+
     private var initialized: Boolean = false
-    private var allowPictureInPictureByCaller = false
+//    private var allowPictureInPictureByCaller = false
     lateinit var workDirectory: File
     var maxBitRate = 705 // k bps
-        private set(v) { field = v }
-    val allowPictureInPicture
-        get() = allowPictureInPictureByCaller && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O
+        private set
+    var allowPictureInPicture = false
+        private set
+    private var httpClientSource: IAmvHttpClientSource? = null
 
+    /**
+     * 初期化. Application.onCreate()から呼び出す。
+     *
+     * @param context   システムがPinPをサポートしているかをチェックするためのcontext (ApplicationContextで可）
+     * @param cacheRootPath ダウンロードした動画のキャッシュ用ディレクトリ
+     * @param bitrate  貼り付ける動画の最大ビットレート
+     * @param allowPinP PinPを許可するか(VfEditionDef の定義）。。。実際にPinPが使えるかどうかは、OSのバージョン＋デバイスの対応状況によって変わる。
+     * @param　httpClientSource　videoライブラリと本体とで、httpClient（の設定やセッションなど）を共用するためのi/f
+     */
     @JvmStatic
-    fun initialize(cacheRootPath: File, bitrate:Int, allowPinP:Boolean) {
+    fun initialize(context: Context, cacheRootPath: File, bitrate:Int, allowPinP:Boolean, httpClientSource:IAmvHttpClientSource?) {
         if (initialized) {
             return
         }
         initialized = true
         maxBitRate = bitrate
-        allowPictureInPictureByCaller = allowPinP
+        allowPictureInPicture = allowPinP && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && context.packageManager.hasSystemFeature(PackageManager.FEATURE_PICTURE_IN_PICTURE)
+        this.httpClientSource = httpClientSource
         val videoCache = File(cacheRootPath, ".video-cache")
         AmvCacheManager.initialize(videoCache)
 
@@ -27,10 +48,14 @@ object AmvSettings {
         if (!workDirectory.exists()) {
             workDirectory.mkdir()
         } else {
-            for (file in workDirectory.listFiles())
-                if (!file.isDirectory) {
+            val files = workDirectory.listFiles() ?: emptyArray()
+            for (file in files)
+                if (null!=file && !file.isDirectory) {
                     file.delete()
                 }
         }
     }
+
+    val httpClient: OkHttpClient
+        get() = httpClientSource?.getHttpClient() ?: OkHttpClient()
 }
